fixing signed/unsigned warning, mentioning T.85 in CHANGELOG and ANNOUNCE

+++ libjbig/jbig85.h
@@ -76,7 +76,7 @@ struct jbg85_enc_state {
   unsigned long y;                       /* next line number to be encoded */
   unsigned long i;            /* next per-stripe line number to be encoded */
   int tx;                           /* x-offset of adaptive template pixel */
-  int c_all, c[128];              /* adaptive template algorithm variables */
+  unsigned long c_all, c[128];    /* adaptive template algorithm variables */
   int new_tx;            /* -1 = no ATMOVE pending, otherwise new TX value */
   int ltp_old;                           /* true if line y-1 was "typical" */
   struct jbg_arenc_state s;                   /* arithmetic encoder status */
